Reeshal

+++ admin/adminhomepage.php
@@ -16,11 +16,14 @@ if ($_SERVER["REQUEST_METHOD"] == "POST"){
         $pid=$_POST['prodId'];
         $email=$_POST['email'];
         if($deleteConfirmation=="yes"){
-            mail($email, "Why we deleted your product", $reason, "From: reeshalnew@gmail.com");
-
-            $delQuery="DELETE FROM Product WHERE productId=$pid";
-            $delResult =$conn->exec($delQuery) ;
-            header("Location: adminhomepage.php?referer=login");
+            if(mail($email, "Why we deleted your product", $reason)){
+                $delQuery="DELETE FROM Product WHERE productId=$pid";
+                $delResult =$conn->exec($delQuery) ;
+                header("Location: adminhomepage.php?referer=login");
+            }
+            else{
+                echo "<script>alert('Email not sent');</script>";
+            }    
         }
         else{
             header("Location: adminhomepage.php?referer=login");
@@ -99,6 +102,10 @@ if ($_SERVER["REQUEST_METHOD"] == "POST"){
                 $("div#viewfaqs").fadeOut(); 
                 $("div#viewusers").fadeOut(); 
                 $("div#viewreports").fadeOut(); 
+                $("#viewchart").removeClass("active");
+                $("#viewuser").removeClass("active");
+                $("#viewfaq").removeClass("active");
+                $("#viewreport").removeClass("active");
              })
 
              $("#viewchart").click(function(){
@@ -108,6 +115,10 @@ if ($_SERVER["REQUEST_METHOD"] == "POST"){
                 $("div#viewfaqs").fadeOut(); 
                 $("div#viewusers").fadeOut(); 
                 $("div#viewreports").fadeOut();
+                $("#viewproduct").removeClass("active");
+                $("#viewuser").removeClass("active");
+                $("#viewfaq").removeClass("active");
+                $("#viewreport").removeClass("active");
              })
 
              $("#viewuser").click(function(){
@@ -116,6 +127,10 @@ if ($_SERVER["REQUEST_METHOD"] == "POST"){
                 $("div#viewreports").fadeOut();
                 $("#viewuser").addClass("active");
                 $("div#viewusers").css("display","block");
+                $("#viewproduct").removeClass("active");
+                $("#viewchart").removeClass("active");
+                $("#viewfaq").removeClass("active");
+                $("#viewreport").removeClass("active");
              })
 
              $("#viewfaq").click(function(){
@@ -124,6 +139,10 @@ if ($_SERVER["REQUEST_METHOD"] == "POST"){
                 $("div#viewreports").fadeOut();
                 $("#viewfaq").addClass("active");
                 $("div#viewfaqs").css("display","block");
+                $("#viewproduct").removeClass("active");
+                $("#viewchart").removeClass("active");
+                $("#viewuser").removeClass("active");
+                $("#viewreport").removeClass("active");
              })
 
              $("#viewreport").click(function(){
@@ -132,6 +151,10 @@ if ($_SERVER["REQUEST_METHOD"] == "POST"){
                 $("div#viewfaqs").fadeOut();
                 $("#viewreport").addClass("active");
                 $("div#viewreports").css("display","block");
+                $("#viewproduct").removeClass("active");
+                $("#viewchart").removeClass("active");
+                $("#viewuser").removeClass("active");
+                $("#viewfaq").removeClass("active");
              })
 
              $('.deletequestion').click(function(){
